Combine all rocket tests in one file
